pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "resend>=2.0.0",
    "supabase>=2.0.0",
]
